inital commit

+++ src/migrations/2017_08_13_185658_create_otp_logs_table.php
@@ -15,7 +15,7 @@ class CreateOtpLogsTable extends Migration
     {
         Schema::create('one_time_password_logs', function (Blueprint $table) {
             $table->increments('id');
-            $table->unsignedInteger("user_id")->index();
+            $table->unsignedBigInteger("user_id")->index();
             $table->string('otp_code')->index();
             $table->string('refer_number')->index();
             $table->string('status')->index();
