+++ src/migrations/2017_08_13_185658_create_otp_table.php
@@ -15,7 +15,7 @@ class CreateOtpTable extends Migration
     {
         Schema::create('one_time_passwords', function (Blueprint $table) {
             $table->increments('id');
-            $table->unsignedInteger("user_id")->index();
+            $table->unsignedBigInteger("user_id")->index();
             $table->string('status')->index();
             $table->timestamps();
         });
